Associate an article with a user.

+++ blog_router/articleHandler.go
@@ -3,6 +3,7 @@ package blog_router
 import (
 	"log"
 	"net/http"
+	"strconv"
 
 	"github.com/gin-contrib/sessions"
 	"github.com/gin-gonic/gin"
@@ -15,6 +16,7 @@ type Article struct {
 	gorm.Model
 	Title string `gorm:"type:string;not null;<-"`
 	Body string `gorm:"type:string;<-"`
+	UserID uint
 }
 
 // データベースを自動的にマイグレーションする
@@ -27,7 +29,7 @@ func Index(c *gin.Context) {
 	if isLogin(c) {
 		username := sessions.Default(c).Get(userKey)
 		articlelist := []Article{}
-		utility.Db.Find(&articlelist)
+		utility.Db.Where("user_id = ?", getUser(username.(string)).ID).Find(&articlelist)
 		c.HTML(http.StatusOK, "article/index.html", gin.H{
 			"articlelist": articlelist,
 			"username": username,
@@ -37,52 +39,67 @@ func Index(c *gin.Context) {
 
 // 記事の詳細表示
 func Show(c *gin.Context) {
-	article := Article{}
-	id := c.Param("id")
-	utility.Db.First(&article, id)
-	c.HTML(http.StatusOK, "article/show.html", article)
+	if isLogin(c) {
+		article := Article{}
+		id := c.Param("id")
+		utility.Db.First(&article, id)
+		c.HTML(http.StatusOK, "article/show.html", article)
+	}
 }
 
 // 記事作成フォームを表示
 func New(c *gin.Context) {
-	c.HTML(http.StatusOK, "article/new.html", nil)
+	if isLogin(c) {
+		username := sessions.Default(c).Get(userKey)
+		user := getUser(username.(string))
+		c.HTML(http.StatusOK, "article/new.html", user)
+	}
 }
 
 // 記事を作成する
 func Create(c *gin.Context) {
-	if err := c.Request.ParseForm(); err != nil {
-		log.Fatalf("フォームの送信に失敗しました: %v", err)
+	if isLogin(c) {
+		if err := c.Request.ParseForm(); err != nil {
+			log.Fatalf("フォームの送信に失敗しました: %v", err)
+		}
+		userid, _ := strconv.Atoi(c.PostForm("userid"))
+		article := Article{Title: c.PostForm("title"), Body: c.PostForm("body"), UserID: uint(userid)}
+		utility.Db.Create(&article)
+		c.Redirect(http.StatusSeeOther, "/article")
 	}
-	article := Article{Title: c.PostForm("title"), Body: c.PostForm("body")}
-	utility.Db.Create(&article)
-	c.Redirect(http.StatusSeeOther, "/article")
 }
 
 // 記事編集フォームを表示
 func Edit(c *gin.Context) {
-	article := Article{}
-	id := c.Param("id")
-	utility.Db.First(&article, id)
-	c.HTML(http.StatusOK, "article/edit.html", article)
+	if isLogin(c) {
+		article := Article{}
+		id := c.Param("id")
+		utility.Db.First(&article, id)
+		c.HTML(http.StatusOK, "article/edit.html", article)
+	}
 }
 
 // 記事を更新する
 func Update(c *gin.Context) {
-	if err := c.Request.ParseForm(); err != nil {
-		log.Fatalf("フォームの送信に失敗しました: %v", err)
+	if isLogin(c) {
+		if err := c.Request.ParseForm(); err != nil {
+			log.Fatalf("フォームの送信に失敗しました: %v", err)
+		}
+		article := Article{}
+		id := c.Param("id")
+		utility.Db.First(&article, id)
+		utility.Db.Model(&article).Updates(Article{Title: c.PostForm("title"), Body: c.PostForm("body")})
+		c.Redirect(http.StatusSeeOther, "/article/" + id)
 	}
-	article := Article{}
-	id := c.Param("id")
-	utility.Db.First(&article, id)
-	utility.Db.Model(&article).Updates(Article{Title: c.PostForm("title"), Body: c.PostForm("body")})
-	c.Redirect(http.StatusSeeOther, "/article/" + id)
 }
 
 // 記事を削除する
 func Destroy(c *gin.Context) {
-	atricle := Article{}
-	id := c.Param("id")
-	utility.Db.First(&atricle, id)
-	utility.Db.Delete(&atricle)
-	c.Redirect(http.StatusSeeOther, "/article")
+	if isLogin(c) {
+		atricle := Article{}
+		id := c.Param("id")
+		utility.Db.First(&atricle, id)
+		utility.Db.Delete(&atricle)
+		c.Redirect(http.StatusSeeOther, "/article")
+	}
 }
